requirements.txt
# Core Deep Learning
torch==2.0.1
torchvision==0.15.2
tensorrt==8.6.1
onnx==1.14.0
onnxruntime-gpu==1.15.1

# Computer Vision
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
ultralytics==8.0.147  # YOLOv8
scikit-image==0.21.0
albumentations==1.3.1
Pillow==10.0.0

# Tracking
filterpy==1.4.5  # Kalman filter
scipy==1.11.1
lap==0.4.0  # Linear assignment

# Graph Neural Networks
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17

# Transformers
transformers==4.31.0
einops==0.6.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
h5py==3.9.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
tensorboard==2.13.0
wandb==0.15.8

# Video Processing
ffmpeg-python==0.2.0
imageio==2.31.1
imageio-ffmpeg==0.4.8

# API & Web
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic==2.1.1
python-multipart==0.0.6
aiohttp==3.8.5
websockets==11.0.3

# Streaming & Messaging
kafka-python==2.0.2
paho-mqtt==1.6.1
redis==4.6.0
celery==5.3.1

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
pymongo==4.4.1

# Monitoring & Logging
prometheus-client==0.17.1
structlog==23.1.0
python-json-logger==2.0.7

# Deployment
docker==6.1.3
kubernetes==27.2.0
boto3==1.28.25  # AWS SDK
azure-identity==1.13.0
google-cloud-storage==2.10.0

# Optimization
onnx-simplifier==0.4.33
torch-pruning==1.2.2

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
click==8.1.6
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
hypothesis==6.82.6

# Explainability
shap==0.42.1
lime==0.2.0.1
captum==0.6.0
grad-cam==1.4.8

# Audio Processing
librosa==0.10.0.post2
soundfile==0.12.1
audioread==3.0.0

# Synthetic Data
carla==0.9.14  # CARLA simulator
gymnasium==0.29.0

# ML Utilities
scikit-learn==1.3.0
imbalanced-learn==0.11.0
optuna==3.3.0  # Hyperparameter tuning

# Privacy & Security
cryptography==41.0.3
pycryptodome==3.18.0

# Development
black==23.7.0
flake8==6.1.0
isort==5.12.0
mypy==1.4.1
pre-commit==3.3.3

# Documentation
sphinx==7.1.2
sphinx-rtd-theme==1.3.0
mkdocs==1.5.2
mkdocs-material==9.1.21

# Jupyter
jupyter==1.0.0
jupyterlab==4.0.5
ipywidgets==8.1.0
notebook==7.0.2

# CUDA utilities (optional, install if using NVIDIA GPUs)
# nvidia-tensorrt==8.6.1
# pycuda==2022.2.2

# Edge-specific (Jetson)
# jetson-stats==4.2.3
